# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ArtHub API",
    "description": "Complete API documentation for ArtHub backend services",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Main API server"
    }
  ],
...
  "paths": {
    "/api/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "الحصول على إشعارات المستخدم",
        "description": "استرجاع جميع إشعارات المستخدم المسجل الدخول",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "رقم الصفحة"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "عدد العناصر في الصفحة"
          }
        ],
        "responses": {
          "200": {
            "description": "تم جلب الإشعارات بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "حذف جميع الإشعارات",
        "description": "حذف جميع إشعارات المستخدم",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "تم حذف جميع الإشعارات بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "deletedCount": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/{notificationId}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "حذف إشعار",
        "description": "حذف إشعار معين",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الإشعار"
          }
        ],
        "responses": {
          "200": {
            "description": "تم حذف الإشعار بنجاح"
          },
          "404": {
            "description": "الإشعار غير موجود"
          }
        }
      }
    },
    "/api/notifications/{notificationId}/read": {
      "patch": {
        "tags": [
          "Notifications"
        ],
        "summary": "وضع علامة \"مقروء\" على إشعار",
        "description": "تعيين إشعار معين كمقروء",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف الإشعار"
          }
        ],
        "responses": {
          "200": {
            "description": "تم تحديث حالة قراءة الإشعار بنجاح"
          },
          "404": {
            "description": "الإشعار غير موجود"
          }
        }
      }
    },
...
    "/api/notifications/read-all": {
      "patch": {
        "tags": [
          "Notifications"
        ],
        "summary": "وضع علامة \"مقروء\" على جميع الإشعارات",
        "description": "تعيين جميع إشعارات المستخدم كمقروءة",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "تم تحديث حالة قراءة جميع الإشعارات بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "modifiedCount": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "job": {
            "type": "string"
          },
          "profileImage": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string",
                "format": "uri"
              }
            }
          }
        }
      },
...
      "NotificationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "notifications": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Notification"
                }
              },
              "pagination": {
                "type": "object",
                "properties": {
                  "currentPage": {
                    "type": "integer"
                  },
                  "totalPages": {
                    "type": "integer"
                  },
                  "totalItems": {
                    "type": "integer"
                  },
                  "itemsPerPage": {
                    "type": "integer"
                  },
                  "unreadCount": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      },
      "Notification": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "sender": {
            "$ref": "#/components/schemas/User"
          },
          "recipient": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}